+++ constants/categoryEndpointMap.ts
